update layout selector, use default OHIF config
