+++ platform/viewer/src/index_module.js
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect } from 'react';
+import React from 'react';
 
 import OHIFDefaultExtension from '@ohif/extension-default';
 import OHIFCornerstoneExtension from '@ohif/extension-cornerstone';
@@ -6,20 +6,46 @@ import OHIFMeasurementTrackingExtension from '@ohif/extension-measurement-tracki
 import OHIFDICOMSRExtension from '@ohif/extension-dicom-sr';
 
 import App from './App.jsx';
-import appConfig from '../public/config/newlantern';
+import defaultConfig from '../public/config/default';
 
-const appProps = {
-  config: appConfig,
-  defaultExtensions: [
-    OHIFDefaultExtension,
-    OHIFCornerstoneExtension,
-    OHIFMeasurementTrackingExtension,
-    OHIFDICOMSRExtension,
-  ],
+const LanternViewer = ({ accessToken, studyUID, gcpDicomURL }) => {
+  const appConfig = {
+    ...defaultConfig,
+    whiteLabeling: {
+      createLogoComponentFn: () => <div></div>,
+    },
+    dataSources: [
+      {
+        friendlyName: 'New Lantern DICOM server',
+        namespace: 'org.ohif.default.dataSourcesModule.dicomweb',
+        sourceName: 'dicomweb',
+        configuration: {
+          name: 'New Lantern',
+          wadoUriRoot: gcpDicomURL,
+          qidoRoot: gcpDicomURL,
+          wadoRoot: gcpDicomURL,
+          qidoSupportsIncludeField: true,
+          imageRendering: 'wadors',
+          thumbnailRendering: 'wadors',
+          enableStudyLazyLoad: true,
+          supportsFuzzyMatching: true,
+          supportsWildcard: false,
+        },
+      },
+    ],
+  };
+
+  const appProps = {
+    config: appConfig,
+    defaultExtensions: [
+      OHIFDefaultExtension,
+      OHIFCornerstoneExtension,
+      OHIFMeasurementTrackingExtension,
+      OHIFDICOMSRExtension,
+    ],
+  };
+
+  return <App {...appProps} accessToken={accessToken} studyUID={studyUID} />;
 };
-
-const LanternViewer = ({ accessToken, studyUID }) => (
-  <App {...appProps} accessToken={accessToken} studyUID={studyUID} />
-);
 
 export default LanternViewer;
